requirements.txt
bardapi>=0.1.30
python-dotenv>=1.0.0
